# models/model_quote.go
/*
 * TestAPI
 *
 * Testing viability of OpenAPI2.0 for new URY API
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type Quote struct {

	QuoteId int32 `json:"quote_id"`

	Text string `json:"text"`

	Source int32 `json:"source"`

	Date string `json:"date"`

	Suspended bool `json:"suspended"`
}